requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohappyeyeballs=2.4.4=py39hecd8cb5_0
aiohttp=3.11.10=py39h46256e1_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=4.7.0=py39hecd8cb5_0
aom=3.6.0=hcec6c5f_0
arrow-cpp=19.0.0=hb704434_4
async-timeout=5.0.1=py39hecd8cb5_0
attrs=24.3.0=py39hecd8cb5_0
aws-c-auth=0.9.0=h46256e1_1
aws-c-cal=0.9.2=h46256e1_0
aws-c-common=0.12.3=h46256e1_0
aws-c-compression=0.3.1=h46256e1_1
aws-c-event-stream=0.5.5=h46256e1_0
aws-c-http=0.10.2=h46256e1_0
aws-c-io=0.20.1=h46256e1_0
aws-c-mqtt=0.13.2=he61ece7_0
aws-c-s3=0.8.3=h46256e1_0
aws-c-sdkutils=0.2.4=h46256e1_0
aws-checksums=0.2.7=h46256e1_0
aws-crt-cpp=0.32.10=h6d0c2b6_0
aws-sdk-cpp=1.11.528=he026a02_1
beautifulsoup4=4.13.4=py39hecd8cb5_0
blas=1.0=openblas
boost-cpp=1.82.0=ha357a0b_2
bottleneck=1.4.2=py39hf3fd67a_0
brotlicffi=1.0.9.2=py39h6d0c2b6_1
bzip2=1.0.8=h6c40b1e_6
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2025.7.15=hecd8cb5_0
cairo=1.16.0=h08824b9_6
certifi=2025.8.3=py39hecd8cb5_0
cffi=1.17.1=py39h9205ec4_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
datasets=2.12.0=py39hecd8cb5_0
dav1d=1.2.1=h6c40b1e_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.13=py39hecd8cb5_0
dill=0.3.6=py39hecd8cb5_0
exceptiongroup=1.2.0=py39hecd8cb5_0
expat=2.7.1=h6d0c2b6_0
filelock=3.17.0=py39hecd8cb5_0
fontconfig=2.14.1=h269ac6a_3
fonttools=4.59.2=pypi_0
freetype=2.13.3=h02243ff_0
fribidi=1.0.10=haf1e3a3_0
frozenlist=1.5.0=py39h46256e1_0
fsspec=2025.7.0=py39he13d532_0
gensim=4.3.3=pypi_0
gettext=0.21.0=h4e8c18a_2
gflags=2.2.2=hcec6c5f_1
glog=0.5.0=hcec6c5f_1
gmp=6.3.0=h6d0c2b6_0
gmpy2=2.2.1=py39h2cec913_0
graphite2=1.3.14=he9d5cce_1
h11=0.16.0=py39hecd8cb5_0
harfbuzz=10.2.0=h060d35a_1
httpcore=1.0.9=py39hecd8cb5_0
httpx=0.28.1=py39hecd8cb5_0
huggingface_hub=0.31.2=py39hecd8cb5_0
icu=73.1=hcec6c5f_0
idna=3.7=py39hecd8cb5_0
importlib-metadata=8.5.0=py39hecd8cb5_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=py39hecd8cb5_0
joblib=1.5.1=py39hecd8cb5_0
jpeg=9e=h46256e1_3
kiwisolver=1.4.7=pypi_0
lcms2=2.16=h31d93a5_1
lerc=4.0.0=h6d0c2b6_0
libabseil=20250127.0=cxx17_h548131f_0
libavif=1.1.1=h46256e1_0
libboost=1.82.0=hf53b9f2_2
libbrotlicommon=1.0.9=h46256e1_9
libbrotlidec=1.0.9=h46256e1_9
libbrotlienc=1.0.9=h46256e1_9
libcurl=8.14.1=h1448931_1
libcxx=19.1.7=haebbb44_3
libdeflate=1.22=h46256e1_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h04015c4_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libglib=2.84.2=heb90364_0
libgrpc=1.71.0=hf2926fd_0
libiconv=1.16=h6c40b1e_3
libnghttp2=1.57.0=h9beae6a_0
libopenblas=0.3.29=h3f8574a_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=5.29.3=h66a43dd_1
libre2-11=2024.07.02=hab2016f_0
libssh2=1.11.1=h3a17b82_0
libthrift=0.15.0=h70b4b81_2
libtiff=4.7.0=h2dfa3ea_0
libtorch=2.6.0=cpu_openblas_h4faba3d_6
libuv=1.48.0=h46256e1_0
libwebp-base=1.3.2=h46256e1_1
libxml2=2.13.8=h6070cd6_0
llvm-openmp=19.1.7=h5048363_2
llvmlite=0.43.0=pypi_0
lz4-c=1.9.4=hcec6c5f_1
markupsafe=3.0.2=py39h46256e1_0
matplotlib=3.9.4=pypi_0
mpc=1.3.1=h46256e1_0
mpfr=4.2.1=h46256e1_0
mpmath=1.3.0=py39hecd8cb5_0
multidict=6.1.0=py39h46256e1_0
multiprocess=0.70.14=py39hecd8cb5_0
ncurses=6.5=h923df54_0
networkx=3.2.1=py39hecd8cb5_0
node2vec=0.5.0=pypi_0
nomkl=3.0=0
numba=0.60.0=pypi_0
numexpr=2.10.1=py39hc59c7be_0
numpy=1.26.4=py39hf6dca73_0
numpy-base=1.26.4=py39hd8f4981_0
oauthlib=3.3.1=py39hecd8cb5_0
odfpy=1.4.1=pypi_0
openjpeg=2.5.2=h2d09ccc_1
openssl=3.0.17=hee2dfae_0
opentelemetry-api=1.30.0=py39hecd8cb5_0
orc=2.1.1=h4dda5a6_0
packaging=25.0=py39hecd8cb5_0
pandas=2.3.1=py39hebe84f7_0
pcre2=10.42=h9b97e30_1
pillow=11.3.0=py39h25d8182_0
pip=25.1=pyhc872135_2
pixman=0.46.4=hb254d66_0
propcache=0.3.1=py39h46256e1_0
protobuf=5.29.3=py39h8d5cf9e_0
pyarrow=19.0.0=py39h6d0c2b6_1
pybind11-abi=4=hd3eb1b0_1
pycparser=2.21=pyhd3eb1b0_0
pyjsparser=2.7.1=py39hecd8cb5_0
pynndescent=0.5.13=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.23=hd8516d5_0
python-dateutil=2.9.0.post0=py39hecd8cb5_1
python-tzdata=2025.2=pyhd3eb1b0_0
python-xxhash=3.5.0=py39h46256e1_0
pytorch=2.6.0=cpu_openblas_py39hfcfd596_6
pytz=2025.2=py39hecd8cb5_0
pyyaml=6.0.2=py39h46256e1_0
re2=2024.07.02=h1962661_0
readline=8.3=h49f2429_0
regex=2024.11.6=py39h46256e1_0
requests=2.32.4=py39hecd8cb5_0
requests-oauthlib=2.0.0=py39hecd8cb5_0
responses=0.13.3=pyhd3eb1b0_0
safetensors=0.5.3=py39hc8aab72_0
scikit-learn=1.6.1=py39h6d0c2b6_0
scipy=1.13.1=py39hb060737_1
sentence-transformers=4.1.0=py39hecd8cb5_0
setuptools=78.1.1=py39hecd8cb5_0
six=1.17.0=py39hecd8cb5_0
sleef=3.5.1=h6c40b1e_2
smart-open=7.3.0.post1=pypi_0
snappy=1.2.1=h6d0c2b6_0
sniffio=1.3.0=py39hecd8cb5_0
soupsieve=2.5=py39hecd8cb5_0
sqlite=3.50.2=hc8b0dd6_1
sympy=1.13.3=py39hecd8cb5_1
tabulate=0.9.0=pypi_0
threadpoolctl=3.5.0=py39h20db666_0
tk=8.6.15=h3a5a201_0
tokenizers=0.21.0=py39h753fa65_0
tqdm=4.67.1=py39h20db666_0
transformers=4.49.0=py39hecd8cb5_0
typing-extensions=4.12.2=py39hecd8cb5_0
typing_extensions=4.12.2=py39hecd8cb5_0
tzdata=2025b=h04d1e81_0
umap-learn=0.5.9.post2=pypi_0
urllib3=2.5.0=py39hecd8cb5_0
utf8proc=2.6.1=h6c40b1e_1
wheel=0.45.1=py39hecd8cb5_0
wrapt=1.17.0=py39h46256e1_0
xxhash=0.8.0=h9ed2024_3
xz=5.6.4=h46256e1_1
yaml=0.2.5=haf1e3a3_0
yarl=1.18.0=py39h46256e1_0
zipp=3.21.0=py39hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
